move TestConnection to rpc_connection
